requirements.txt
torch~=2.5.1
torchvision~=0.20.1
opencv-python~=4.11.0.86
numpy~=2.0.1
matplotlib~=3.10.0
pillow~=11.1.0
PyYAML~=6.0.2
scipy~=1.15.1
imageio~=2.37.0
scikit-image~=0.25.1
lpips~=0.1.4
openai~=1.65.1
sentence-transformers~=3.4.1